api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.17.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
async-timeout==5.0.1
    # via asyncpg
asyncpg==0.30.0
    # via -r requirements.in
exceptiongroup==1.3.0
    # via anyio
fastapi==0.119.0
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
idna==3.11
    # via anyio
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
pydantic==2.12.2
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   alembic
starlette==0.48.0
    # via fastapi
tomli==2.3.0
    # via alembic
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings

uvicorn[standard]==0.23.2